+++ pywander/__init__.py
@@ -5,6 +5,6 @@
 a general python module
 """
 
-__version__ = '0.11.1'
+__version__ = '0.12.0'